feat: rework various components
