+++ posts/2020_06_16_driving_stepper_motors_with_microcontroller.md
@@ -35,7 +35,7 @@ Also included
 - [12v Wall wart](https://smile.amazon.com/dp/B07DCPT1N7/ref=cm_sw_em_r_mt_dp_U_dz28EbW3A14S0)
 - [Barrel jack plugs](https://smile.amazon.com/dp/B074LK7G86/ref=cm_sw_em_r_mt_dp_U_Rz28Eb3CF0KZ4)
 
-![A drv8833 stepper motor in a breakout board format.](images/2020_06_16_driving_stepper_motors_with_microcontroller/drv8833.jpeg)
+![A drv8833 stepper motor in a breakout board format.](/images/post/2020_06_16_driving_stepper_motors_with_microcontroller/drv8833.jpeg)
 
 This drv8833 I had to solder the header's to it with the help of my friend Johhny.
 
@@ -49,7 +49,7 @@ Typically the number designates the faceplate size of the motor
 eg the four mounting points around the shaft, you can typically get the same size motor in different lengths for more or less torque. The
 NEMA 17 for instance has 1.7" inch faceplates. Below is a bunch of different sized images.
 
-![Diffrent sized NEMA motors](images/2020_06_16_driving_stepper_motors_with_microcontroller/nema-stepper-motors.jpg)
+![Diffrent sized NEMA motors](/images/post/2020_06_16_driving_stepper_motors_with_microcontroller/nema-stepper-motors.jpg)
 
 ### Power
