+++ src/pages/posts/[slug].tsx
@@ -1,7 +1,11 @@
+import CodeBlock from "@components/CodeBlock";
 import PageLayout from "@components/PageLayout";
-import { Post, getContent } from "@utils/markdown";
+import { Post } from "@utils/markdown";
+import { Language } from "prism-react-renderer";
 import React from "react";
 import { GetStaticPaths, GetStaticProps } from "next";
+import { Remark } from "react-remark";
+import Image from "next/image";
 
 const PostPage: React.FC<Post> = (props) => {
   return (
@@ -10,12 +14,62 @@ const PostPage: React.FC<Post> = (props) => {
         <div className="relative px-4 sm:px-6 lg:px-8">
           <div className="text-lg max-w-prose mx-auto">
             <h1>
-              <span className="block text-base text-center text-indigo-600 font-semibold tracking-wide uppercase">
+              <span className="block text-base text-center text-indigo-600 font-semibold tracking-wide uppercase pb-6">
                 {props.title}
               </span>
             </h1>
           </div>
-          <div>{getContent(props.content)}</div>
+          {/*<div>{getContent(props.content)}</div>*/}
+          <div className="flex flex-col space-y-3 text-base max-w-prose mx-auto lg:max-w-none">
+            <Remark
+              rehypeReactOptions={{
+                components: {
+                  pre: (props) => {
+                    const className: string =
+                      props.children[0].props?.className;
+                    if (className && className.includes("language-")) {
+                      let language = className.replace("language-", "");
+                      const code = props.children[0].props.children[0];
+                      if (language === "typescript") {
+                        language = "jsx";
+                      }
+
+                      return (
+                        <CodeBlock
+                          code={code}
+                          language={language as Language}
+                        />
+                      );
+                    }
+
+                    return <pre>{props.children}</pre>;
+                  },
+                  // img: (props) => (
+                  //   <Image
+                  //     className=" flex justify-centermax-w-xs max-h-xs"
+                  //     height={200}
+                  //     width={400}
+                  //     src={props.src as string}
+                  //     alt={props.alt as string}
+                  //   />
+                  // ),
+                  p: (props) => {
+                    console.log(props);
+                    return (
+                      <div className="">
+                        <p className="">{props.children}</p>
+                      </div>
+                    );
+                  },
+                  ul: (props) => (
+                    <ul className="list-disc list-inside">{props.children}</ul>
+                  ),
+                },
+              }}
+            >
+              {props.content}
+            </Remark>
+          </div>
         </div>
       </div>
     </PageLayout>
