+++ src/utils/markdown.tsx
@@ -1,10 +1,5 @@
-import CodeBlock from "@components/CodeBlock";
-import { Language } from "prism-react-renderer";
 import React from "react";
 import fm, { FrontMatterResult } from "front-matter";
-import unified from "unified";
-import parse from "remark-parse";
-import remark2react from "remark-react";
 
 interface PostAttributes {
   slug: string;
@@ -36,31 +31,3 @@ export const getPosts = (fs): Array<Post> => {
   });
   return paths;
 };
-
-export const getContent = (markdown: string) =>
-  unified()
-    .use(parse)
-    .use(remark2react, {
-      sanitize: { attributes: { "*": ["className"] } },
-      remarkReactComponents: {
-        pre: (props) => {
-          const className: string = props.children[0].props?.className;
-          if (className && className.includes("language-")) {
-            let language = className.replace("language-", "");
-            const code = props.children[0].props.children[0];
-            if (language === "typescript") {
-              language = "jsx";
-            }
-
-            return <CodeBlock code={code} language={language as Language} />;
-          }
-
-          return <pre>{props.children}</pre>;
-        },
-        // code: (props) => (
-        //   <code className="overflow-x-auto">{props.children}</code>
-        // ),
-      },
-    })
-    // .use(prism)
-    .processSync(markdown).result;
